chore: 단위, 통합 테스트 환경 분리
