+++ src/test/intg/java/toy/bookswap/BookswapApplicationTests.java
@@ -1,0 +1,15 @@
+package toy.bookswap;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+@ActiveProfiles("intg")
+@SpringBootTest
+public class BookswapApplicationTests {
+
+  @Test
+  void contextLoads() {
+
+  }
+}
