--- src/test/java/toy/bookswap/BookswapBackApplicationTests.java
@@ -1,13 +1,0 @@
-package toy.bookswap;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BookswapBackApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
