+++ src/test/unit/java/toy/bookswap/UnitTest.java
@@ -1,0 +1,14 @@
+package toy.bookswap;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.Test;
+
+public class UnitTest {
+
+  @Test
+  public void test() {
+    assertThat(true).isTrue();
+  }
+}
+
